Initial Fusion service
Fusion is intended to provide an easy way to quickly start/stop multiple services together for testing purposes
