+++ src/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright (c) 2016 - 2021 Ember
+# Copyright (c) 2016 - 2024 Ember
 #
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -13,3 +13,4 @@ add_subdirectory(mdns)
 add_subdirectory(character)
 add_subdirectory(libs)
 add_subdirectory(tools)
+add_subdirectory(fusion)
