+++ src/mdns/CMakeLists.txt
@@ -18,6 +18,7 @@ set(LIBRARY_HDR
 	Serialisation.h
 	detail/Parser.h
 	NSDService.h
+	Runner.h
     )
 
 set(LIBRARY_SRC
@@ -28,6 +29,7 @@ set(LIBRARY_SRC
 	Serialisation.cpp
 	detail/Parser.cpp
 	NSDService.cpp
+	Runner.cpp
     )
 
 add_library(${LIBRARY_NAME} ${LIBRARY_HDR} ${LIBRARY_SRC})
